comment in the index of events and highlight the box of comments

+++ lib/tasks/marte.rake
@@ -7,7 +7,9 @@ namespace :marte do
       @past_events.select{|e| 
 	begin
 	  room = MarteRoom.find(e.id)
-	  room.destroy
+	  if room
+		room.destroy
+	  end
 	  e.marte_room = 0
 	rescue
 	  #if the room does not exist, delete it from the event
